Fixed issue with dollar sign. Updates according to PR comments.
