+++ btr-web/btr-common-components/pages/examples/payWidget.vue
@@ -14,20 +14,20 @@ const fees: PayFeesWidgetItemI[] = [
     processingFees: 0,
     serviceFees: 0,
     tax: { gst: 0, pst: 0 },
-    total: 0
+    total: 12
   },
   {
     uiUuid: '123',
     quantity: 2,
     filingFees: 0,
     filingType: 'Register Significant Individual2',
-    filingTypeCode: 'REGSIGIN2',
+    filingTypeCode: 'TEST',
     futureEffectiveFees: 0,
     priorityFees: 0,
     processingFees: 0,
     serviceFees: 0,
     tax: { gst: 0, pst: 0 },
-    total: 0
+    total: 27.23
   }
 ]
