+++ btr-web/btr-main-app/app/router.options.ts
@@ -40,7 +40,7 @@ export default <RouterConfig> {
           leftButtons: [],
           rightButtons: []
         },
-        layout: 'filing'
+        layout: 'business'
       }
     },
     {
