+++ btr-web/devops/vaults.env
@@ -15,8 +15,8 @@ VUE_APP_REGISTRY_HOME_URL="op://web-url/$APP_ENV/registry/REGISTRY_HOME_URL"
 VUE_APP_AUTH_WEB_URL="op://web-url/$APP_ENV/auth-web/AUTH_WEB_URL"
 
 #vaults API
-#VUE_APP_PAY_API_URL="op://API/$APP_ENV/pay-api/PAY_API_URL"
-#VUE_APP_PAY_API_VERSION="op://API/$APP_ENV/pay-api/PAY_API_URL  "
+VUE_APP_PAY_API_URL="op://API/$APP_ENV/pay-api/PAY_API_URL"
+VUE_APP_PAY_API_VERSION="op://API/$APP_ENV/pay-api/PAY_API_URL  "
 VUE_APP_AUTH_API_URL="op://API/$APP_ENV/auth-api/AUTH_API_URL"
 VUE_APP_AUTH_API_VERSION="op://API/$APP_ENV/auth-api/AUTH_API_VERSION"
 VUE_APP_LEGAL_API_URL="op://API/$APP_ENV/legal-api/LEGAL_API_URL"
